requirements.txt
streamlit==1.28.1
anthropic==0.7.7
python-dotenv==1.0.0